Missing dependency added
There is a missing library in the libsodium dependencies that caused me trouble with the compiling process when i ran "cabal build cardano-cli cardano-node". Specifically, when it tries to compile things relatives to the cardano-crypto-class library. Further details are on this Cardano forum thread: https://forum.cardano.org/t/when-compiling-fails/85525/5